// client/src/utilities/__generated__/login.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: login
// ====================================================

export interface login {
  loginUser: string;
}

export interface loginVariables {
  username: string;
  email: string;
}